feat: transfer functionalities
