+++ packages/frontend/src/components/wallet-migration/MigrateAccounts.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useCallback, useState} from 'react';
 import { Translate } from 'react-localize-redux';
 import styled from 'styled-components';
